@@ -12,23 +12,22 @@
 		"fideloper/proxy": "^4.4",
 		"fruitcake/laravel-cors": "^2.0",
 		"doctrine/dbal": "2.5.13",
-		"guzzlehttp/guzzle": "^6.0",
+		"guzzlehttp/guzzle": "^7.0.1",
 		"intervention/image": "^2.7",
-		"laravel/framework": "^7.0",
+		"laravel/framework": "^8.0",
 		"laravel/tinker": "^2.0",
 		"maxmind-db/reader": "1.4.1",
 		"namshi/jose": "7.0.*",
 		"symfony/lock": "^4.0",
 		"tymon/jwt-auth": "^1.0",
-		"fideloper/proxy": "^4.0",
-		"laravel/ui": "^2.0"
+		"laravel/ui": "^3.0"
 	},
 	"require-dev": {
-		"facade/ignition": "^2.0",
-		"phpunit/phpunit": "^8.5",
+		"facade/ignition": "^2.3.6",
+		"phpunit/phpunit": "^9.0",
 		"fakerphp/faker": "^1.9.1",
 		"mockery/mockery": "^1.3.1",
-		"nunomaduro/collision": "^4.3"
+		"nunomaduro/collision": "^5.0"
 	},
 	"autoload": {
 		"classmap": [
